move files in DGtal/math
