+++ tests/geometry/CMakeLists.txt
@@ -6,8 +6,7 @@ add_subdirectory(tools)
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O2 -g")
 
 SET(DGTAL_TESTS_SRC 
-  testContourHelper
-  testProfile)
+  testContourHelper)
 
 FOREACH(FILE ${DGTAL_TESTS_SRC})
   add_executable(${FILE} ${FILE})
